controller - create multiple error classifications
Error submodules are preferred for error types, and it is preferable to
create multiple error classifications rather than bubbling all errors up
to a single error type. In this change, we create multiple error
classifications for controller.
