+++ controller/src/lib.rs
@@ -1,7 +1,6 @@
 mod controller;
 mod metrics;
 
-pub mod error;
 pub mod statemachine;
 pub mod telemetry;
